Update: Add house-keeping methods to interact with Drive API
Add a couple of house-keeping methods that will interact with Google
API and setup stuff for the rest of the script(s).

This includes authenticating the connection, and fetching a list of
teamdrives that are linked to the google account - allowing the user
to select one of them to be able to use it as a source to generate
strm files.