Update: Basic version
This commit includes a basic running version of the script.

The script can now traverse a directory structure while generating
the strm files required. Users can freely control the source
folder/drive and the destination where the resulting structure
is being generated.

Optionally, if no source is provided, the script will fetch a list
of all team-drives to which the user has access, and then ask the user
to select one of them as the source directory.