BugFix: Patch a minor bug. Improve documentation.
Fix a minor bug. The bug was caused because the script assumed that
the source of all files being scanned was a teamdrive, due to this,
incase of files present in 'My Drive', the script attempted to
access non-existant fields (which would have otherwise been present
in an item stored inside a teamdrive). This led to a crash mid-way
if a scan was being performed on an item present in normal Drive
storage.

Improve the documentation to keep it up-to-date with the changes
being made in this commit. Change a few existing bits to fix
typo's, and just dumb it down in general.

Make sure that the script wipes the existing destination directory
before the actual cloning begins. This ensures that items from a
previous scan are not left in the destination drives by the time
the cloning finishes. Minor modification to the script so that
teamdrives too get a directory of their name instead of a simple
'Drive'.