Update: Patch some major bugs. Push updates to the console.
This commit patches a few major bugs. Due to changes between Drive
API v3 and v2, the method being used to fetch the list of teamdrives
(when no console arugment was used with the script) ran into error
with the method that scanned the source(s).

Apart from this, due to the v2 API, files being scanned included
files that had been trashed (deleted) and were not a part of the
source. This commit patches both these bugs.

The script can now run without any command-line arguments,
additionally, files that have been trashed will no longer be a part
of the files being scanned.

Additionally, this commit also pushes live updates to the terminal
as files are being scanned - this is thanks to the `reprint` module.

Obviously, not everyone would want their console to be cluttered with
updates (me for one - I run this script in the background). This update
also includes a flag to silence the script (i.e. no updates will be
pushed to the terminal), simply use the argument `--updates=off`
while running the script to make the script run silently - I'll update
the readme with this info sometime later.