Update: Add info regarding directory being scanned during updates
This commit prints info on the screen regarding the directory being
scanned on drive. The paths being printed are relative to the folder
structure on drive.

Apart from that, this commit also adds a flag to be able to modify
the name of the root directory inside which the strm files are being
generated.