Release: Update readme with changes. Minor patches. Fix typo(s)
Update readme with the changes made in the previous commits - a large
part of the readme being rewritten in a slightly modified convention
in an attempt to make it be more simpler and easier to understand.

Patch up the bat setup file - includes `@echo off` this time :p

Patch a few potential bugs in the main script - includes a slight
modification to the regex pattern being used in a particular case.

Modify gitignore to ignore directories name `venv` or `.venv`.
This commit prepares the repo for a release - the main feature in
this release being the flag to turn console updates on or off, and
being able to rename the root directory being created by the script
during the run (the lack of this being something that has been
bothering me from quite a while).