BugPatch: Fix a minor typo in strm file(s)
This commit includes an otherwise minor bug patch that 
caused the script to break mid-way.

Fix the content of the strm file being written to, to ensure 
that the file can be parsed correctly by the add-on as 
needed. Due to some changes in the add-on itself, 
and a minor typo on my end, the strm files being 
generated could not be parsed correctly by the add-on,
as a result. This change is intended to fix this error.