BugFix: Minor bugfix in contents of the strm file(s). Refactoring.
Should fix up the `play() takes at least 2 arguments (3 given)`
error message that a handful of systems seem to be having while
attempting to play (any) strm file generated by this script.

Turns out the actual reason of the error was pretty stupid -
instead of the strm file containing `driveid` it had `driveId`
(notice the capital `I` instead of small `i`). All problems
caused due to capitalization of one alphabet.

For those interested in the details, turns out what the add-on
(internally) does is - it parses text from the strm files,
breaks it down into key-value pairs, and then attempts to use
this dictionary (of sorts) as *named* parameters being used
in an actual function call. A very stupid way to do this...
And another reason for me to dislike python :(

Since the key-value pairs were being used in a function call,
the capitalization ensured that `driveId` is not the same argument
as `driveid` - which resulted in a missing parameter during the
function call, and thus the error. The change being made in
this commit should be enough to fix it up.

Additionally, this commit also includes a bunch of code
refactoring, includes (finally) creating a main method to avoid
cluttering the global namespace with all the variables. Fix a lot
of other typo's from comments, trim the comments to reduce the fluff
and more.