BugFix: Minor bugfix. Auto-selection of root directory name
Patch a minor bug being caused because the script was unable to
choose a name for the root directory by itself. The script
worked fine if a name for the root directory was being supplied
using CLI, otherwise the script ran into an error as name of
the root directory had not been initialized.