Major: Breaking changes. Modify OAuth Scope. Update readme
This commit prepares the repo for a future commit allowing users
to optionally chose between using a pre-made project for
authentication, or using their own project for this purpose.

As such, this commit modifies the Drive OAuth scope being used
to authenticate connection with Drive. Instead of requiring complete
read and write permission, modify the script to now use just read
permissions - without this users in future could potentially
give write permission to the pre-made Google Project.

As a result, the previous `token.pickle` file needs to be deleted,
and a new one will be created after a login process by this updated
script.

Modify the readme to contain information regarding this update.