Refactor: Convert method to private, and other changes
Convert method `get_teamdrives` to be private method in the class
`DriveHandler` - since the only call to the method was needed
before a call to `DriveHandler.walk`, this commit converts the
method into a private method. The method `DriveHandler.walk`
will internally make a call to the (now) private method

This change simplifies the external API of the class

The instance-variable `DriveHandler.dirs` was modified to cache
directory names, as opposed to caching just teamdrive names

Modify the private method `__get_teamdrives` - originally, the
method would add teamdrives to the instance variable `dirs`. With
this change, the method will return a dictionary of the teamdrives
found, and add the same to the cache `DriveHandler.dirs`